Experimental: TG on JSR (Node.js only, released as 1.4.2)
We've decided to release a very experimental version of TouchGuild, particularly ESM part on JSR, a package manager for modern TypeScript, though, we do not plan, for now, to make our library fully compatible to other runtimes. TouchGuild on JSR may lack some documentation symbols, and you may face issues when using, though this support will probably stay experimental, and issues won't be fixed if they involve significant architectural changes. It basically is TouchGuild published on JSR, so, nothing much.. Being on this package manager can make TouchGuild discovered by Bun users (with npm support, and so, node), but it doesn't mean anything to Deno ones. For better support and maintainability, we highly recommend to switch on Node.js, which can be unfortunate for some people.