Explicit/Positional & Value Transposition for App Cmd
Application Commands now support the explicit form as well as the positional one, allowing optionals & required options to be specified in any order, and both can work at the same time, making things much more flexible for everyone. You can now use Application Commands the way you like the most and that is the most practical, natural and easy to use for you. This drastically enhances the way users interact with commands.